written test cases for login and signup
